+++ libraries/BluetoothSerial/examples/DiscoverConnect/DiscoverConnect.ino
@@ -64,7 +64,7 @@ void setup() {
         BTAdvertisedDevice *device = btDeviceList->getDevice(i);
         Serial.printf(" ----- %s  %s %d\n", device->getAddress().toString().c_str(), device->getName().c_str(), device->getRSSI());
         std::map<int, std::string> channels = SerialBT.getChannels(device->getAddress());
-        Serial.printf("scanned for services, found %zu\n", channels.size());
+        Serial.printf("scanned for services, found %lu\n", (unsigned long)channels.size());
         for (auto const &entry : channels) {
           Serial.printf("     channel %d (%s)\n", entry.first, entry.second.c_str());
         }
