+++ tests/validation/nvs/nvs.ino
@@ -53,7 +53,7 @@ void setup() {
   uint16_t val_ushort = preferences.getUShort("ushort", 0);
   int32_t val_int = preferences.getInt("int", 0);
   uint32_t val_uint = preferences.getUInt("uint", 0);
-  int64_t val_long = preferences.getLong("long", 0);
+  int32_t val_long = preferences.getLong("long", 0);
   uint32_t val_ulong = preferences.getULong("ulong", 0);
   int64_t val_long64 = preferences.getLong64("long64", 0);
   uint64_t val_ulong64 = preferences.getULong64("ulong64", 0);
@@ -77,8 +77,8 @@ void setup() {
   }
 
   Serial.printf("Values from Preferences: ");
-  Serial.printf("char: %c | uchar: %u | short: %d | ushort: %u | int: %ld | uint: %lu | ", val_char, val_uchar, val_short, val_ushort, val_int, val_uint);
-  Serial.printf("long: %lld | ulong: %lu | long64: %lld | ulong64: %llu | ", val_long, val_ulong, val_long64, val_ulong64);
+  Serial.printf("char: %c | uchar: %u | short: %d | ushort: %u | int: %" PRIi32 " | uint: %" PRIu32 " | ", val_char, val_uchar, val_short, val_ushort, val_int, val_uint);
+  Serial.printf("long: %" PRIi32 " | ulong: %" PRIu32 " | long64: %" PRIi64 " | ulong64: %" PRIu64 " | ", val_long, val_ulong, val_long64, val_ulong64);
   Serial.printf(
     "float: %.2f | double: %.2f | bool: %s | str: %s | strLen: %s | struct: {id:%u,val:%u}\n", val_float, val_double, val_bool ? "true" : "false",
     val_string.c_str(), val_string_buf, test_data.id, test_data.value
